+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Alarm plugin for Clock, Weather Station, ...
 - NTP plugin: Check once per minute if the time is valid and call configTime() again if not
+- MQTT metrics: uptime, heap
 
 ## Version 0.0.3